# specs/001-llm-course-select/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: LLM 自动选课系统 API
  version: 0.1.0
  description: AI 驱动的智能选课推荐平台 API 接口规范（系统不执行真实选课操作）
servers:
  - url: http://localhost:8000/api/v1
    description: 开发环境
paths:
...
  /plans:
    post:
      summary: 收藏推荐方案
      description: 将推荐方案保存到"我的方案"（FR-05）
      operationId: savePlan
      tags: [方案]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, course_ids]
              properties:
                name:
                  type: string
                  description: 方案名称
                course_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: 课程 ID 列表
                notes:
                  type: string
                  description: 备注
      responses:
        '201':
          description: 收藏成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedPlan'

    get:
      summary: 获取我的方案列表
      operationId: listPlans
      tags: [方案]
      responses:
        '200':
          description: 方案列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  plans:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavedPlan'
  /plans/{plan_id}/export:
    get:
      summary: 导出方案为 PDF
      description: 将收藏的方案导出为 PDF 文件（FR-05）
      operationId: exportPlan
      tags: [方案]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF 文件
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /plans/{plan_id}:
    delete:
      summary: 删除收藏方案
      operationId: deletePlan
      tags: [方案]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
...
components:
  schemas:
...
    SavedPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        course_ids:
          type: array
          items:
            type: string
            format: uuid
        total_credits:
          type: number
        match_score:
          type: number
        notes:
          type: string
        created_at:
          type: string
          format: date-time
...